% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giro.R
\docType{data}
\name{giro}
\alias{giro}
\title{GC tables for all giro editions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 77 rows and 2 columns.
}
\usage{
giro
}
\description{
GC tables for all giro editions
}
\keyword{datasets}
