% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editions.R
\docType{data}
\name{editions}
\alias{editions}
\title{Information the Giro d'Italia}
\format{

}
\source{
\url{https://www.procyclingstats.com/race/giro-d-italia}
}
\usage{
editions
}
\description{
A dataset containing informations about all the previous editions (from 1909 to 2021)
of the Giro d'Italia, one of the 3 grand tours in cycling.
}
\keyword{datasets}
